requirements.txt.txt
amqp==5.0.2
anyascii==0.2.0
asgiref==3.3.4
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
billiard==3.6.3.0
blis==0.2.4
celery==5.0.5
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.2.1
cymem==2.0.5
dj-database-url==0.5.0
Django==3.1.12
django-celery-beat==2.1.0
django-classy-tags==2.0.0
django-cms==3.8.0
django-colorful==1.3
django-el-pagination==3.3.0
django-filer==2.0.2
django-filter==2.4.0
django-formtools==2.3
django-guardian==2.3.0
django-js-asset==1.2.2
django-microsoft-auth==2.4.0
django-modelcluster==5.1
django-mptt==0.12.0
django-phonenumber-field==5.0.0
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-social-share==2.2.1
django-taggit==1.4.0
django-timezone-field==4.1.1
django-treebeard==4.4
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-installer==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
djangorestframework==3.12.4
draftjs-exporter==2.1.7
easy-thumbnails==2.7.1
et-xmlfile==1.1.0
5lib==1.1
idna==2.10
install==1.3.4
isort==5.6.4
joblib==1.0.1
kombu==5.0.2
l18n==2020.6.1
lazy-object-proxy==1.4.3
mathparse==0.1.2
mccabe==0.6.1
msal==1.8.0
murmurhash==1.0.5
nltk==3.6.2
numpy==1.21.1
oauthlib==3.1.0
openpyxl==3.0.7
packaging==21.0
phonenumbers==8.12.13
Pillow==8.2.0
Pint==0.17
plac==0.9.6
preshed==2.0.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pymongo==3.12.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.7.5
pytz==2020.4
PyYAML==5.1
redis==3.5.3
regex==2021.8.3
requests==2.25.0
requests-oauthlib==1.3.0
sib-api-v3-sdk==7.2.0
six==1.15.0
soupsieve==2.2.1
#spacy==2.1.9
SQLAlchemy==1.2.19
sqlparse==0.4.1
srsly==1.0.5
tablib==3.0.0
telepath==0.2
thinc==7.0.8
toml==0.10.2
tqdm==4.62.0
twilio==6.49.0
tzlocal==2.1
Unidecode==1.1.2
urllib3==1.26.2
vine==5.0.0
wagtail==2.13.1
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Willow==1.4
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0
